refactor: move properties V1 module
