+++ gravitee-apim-console-webui/src/management/api/_api.scss
@@ -3,7 +3,7 @@
 @import 'creation-v2/steps/api-creation';
 @import 'audit/history/apiHistory';
 @import 'policy-studio-v1/policies/policy';
-@import 'proxy/properties-v1/properties';
+@import 'properties-v1/properties';
 @import 'proxy/health-check-dashboard/healthcheck';
 @import 'resources-v1/resources';
