+++ gravitee-apim-console-webui/src/management/api/properties-v1/add-property.dialog.controller.ts
