+++ gravitee-apim-console-webui/src/management/api/properties-v1/dynamic-provider-http.dialog.html
