+++ gravitee-apim-console-webui/src/management/api/properties-v1/properties.component.ajs.ts
