+++ gravitee-apim-console-webui/src/management/api/proxy/api-proxy.module.ts
@@ -17,7 +17,6 @@
 import { NgModule } from '@angular/core';
 
 import { ApiProxyHealthCheckDashboardModule } from './health-check-dashboard/api-proxy-health-check-dashboard.module';
-import { ApiV1PropertiesComponent } from './properties-v1/properties.component';
 import { ApiPropertiesModule } from './properties/properties/api-properties.module';
 import { ApiDynamicPropertiesModule } from './properties/dynamic-properties/api-dynamic-properties.module';
 
@@ -25,7 +24,7 @@ import { ApiV1ResourcesComponent } from '../resources-v1/resources.component';
 import { ApiProxyCorsModule } from '../cors/api-proxy-cors.module';
 
 @NgModule({
-  declarations: [ApiV1PropertiesComponent, ApiV1ResourcesComponent],
+  declarations: [ApiV1ResourcesComponent],
   imports: [ApiProxyCorsModule, ApiProxyHealthCheckDashboardModule, ApiPropertiesModule, ApiDynamicPropertiesModule],
 })
 export class ApiProxyModule {}
